+++ src/secrets/index.ts
@@ -1,3 +1,5 @@
 export type { SecretStore, SecretStoreType } from "./types.js";
 export { createSecretStore } from "./factory.js";
 export { detectStoreType } from "./utils.js";
+export { NoopSecretStore } from "./none.js";
+export { FileSecretStore } from "./file.js";